requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
croniter==0.3.30
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.40.0
django-cors-headers==3.4.0
django-filter==2.2.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-rest-swagger==2.2.0
django-settings-file==2.2.0
djangorestframework==3.9.4
gevent==20.9.0
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.0a1
hvac==0.9.6
idna==2.8
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-client==0.0.1
oauthlib==3.1.0
openapi-codec==1.3.2
orderedset==2.0.3
packaging==20.4
pluggy==0.13.1
promise==2.3
psycopg2-binary==2.8.6
py==1.9.0
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.7.1
pytest-django==3.5.1
python-dateutil==2.8.1
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
Rx==1.6.1
sendgrid==6.6.0
simplejson==3.17.2
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
toml==0.10.2
tornado==6.0.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.11
whitenoise==3.0
zope.event==4.5.0
zope.interface==5.2.0
